feat: Update catalog headers and env vars (#24)
* fix: Update get catalog headers. Update SDK

* fix: Update Config

* fix: Update config variables

* fix: Add .env

* fix: Add next public prefix

* fix: Remove console

* refactor: algolia env values check

* refactor: removed REACT_APP prefix

Not needed with nextjs

* refactor: formatting

* feat: runtime env variable check

Check if algolia env vars have been provided

throw descriptive error if they have not

* feat: runtime env variable check stripe

Check if stripe env vars have been provided

throw descriptive error if they have not

* feat: add ep channels and tags

* fix: corrected name of env variable

* fix: corrected name of env variable

* feat: api endpoint for client side promotion fetch

* refactor: camelcase

* refactor: make it clear that this can only be used server side

* fix: workaround for handling serverless epcc js-sdk auth

Warning! This is not a permanent solution. Every time we hit the server side we will be running auth against epcc.

* refactor: clean up async chaining

* fix: hidden secrets

Secrets are now spicified in netlify

to work locally copy the .env.example and name it .env.local

* refactor: marked required values

* chore: updated readme to explain local env

Co-authored-by: Robert Field <robertfield@quadrimular.com>