feat: product pages now use just the product id in the url
Simplified product page route to only use the product id in the route. Everything else is fetched based on that id.

This id is not human-readable we want to replace it with slugs as possible.

This change allows for easy navigation from Algolia index entries which only have a product id no parent ids